fix bug #0001367: User should have a way to remove pending node and cancel the polling thread.
In PollingThread.java, add two function like this:
	public void cancel(){
		cancelled = true;
	}

	public GridEndpointReferenceType getGridEPR(){
		return gridEPR;
	}

And check variable "cancelled" before getResults().
But we need someone to receive the event, because PollingThread is not a component, so it can not receive event.

I choose AnalysisPanel to receive the PendingNodeCancelledEvent.
    @Subscribe
	public void receive(org.geworkbench.events.PendingNodeCancelledEvent e, Object source){
    	for (Iterator iterator = threadList.iterator(); iterator.hasNext();) {
    		PollingThread element = (PollingThread) iterator.next();
			if (element.getGridEPR()==e.getGridEpr()){
				element.cancel();
			}
		}
	}

Of course, inorder to have a threadList used in above method,
We need to put "threadList.add(pollingThread);" when ever we new a pollingThread. (two places in AnalysisPanel.java)

Then, we need someone to fire the PendingNodeCancelledEvent.
That would be in fileRemove_actionPerformed() in ProjectPanel.java, when user right click on project panel and choose to remove the pending node.

Of course, user need to see the menu to select remove-the-pending-node option.
It's in jProjectTree_mouseReleased() in ProjectPanel.java

Total files modified: Three
1. PollingThread.java, 2. AnalysisPanel.java, 3. ProjectPanel.java
One new file: PendingNodeCancelledEvent.java


git-svn-id: https://ncisvn.nci.nih.gov/svn/geworkbench/trunk/geworkbench@3877 8b29bb6e-589a-435a-9a12-3ab907be3462