to fix 0001507: EVD display and marker sets not correctly linked one file updated.
In receive(GeneSelectorEvent e, Object source) in MicroarrayViewEventBase.java,
I changed
    refreshMaSetView()
To
    if (markers!=null)
        refreshMaSetView();

Because in refreshMaSetView(), variable maSetView and uniqueMarkers will be changed.
But when we single click on a marker in markers panel, we didn't expect to change anything.
So the maSetView shouldn't be changed, nor the uniqueMarkers, nor fireModelChangedEvent().

So I change it to only call refreshMaSetView() when we do really have markers activated or deactivated.
In following cases (activate, deactivate, add marker, remove marker), markers won't be null,
But in case of just select the marker but didn't add to marker set, markers variable will be null.
So I use markers!=null as checking condition.


git-svn-id: https://ncisvn.nci.nih.gov/svn/geworkbench/trunk/geworkbench@4153 8b29bb6e-589a-435a-9a12-3ab907be3462