fix bug 0001523: axis are mis-labeled
The reason is, the code using the index in scatter plot (the list you click to choose markers or arrays to plot) to get the marker or array.
But since the list in scatter plot was already been filtered when activate marker sets or microarray sets; it needs to be map back to the index in our original index in our BISON object.

The mapping mechanism has been implemented,
One file updated on CVS: ScatterPlot.java


git-svn-id: https://ncisvn.nci.nih.gov/svn/geworkbench/trunk/geworkbench@4195 8b29bb6e-589a-435a-9a12-3ab907be3462