Fixed bug#2073, allow "open file" window for local and remote(CaArray) to be adjustable.
Changes made to /src/org/geworkbench/builtin/projects/LoadData.java: In jbInit(), set current JDialog window to be resizable.
Changes made to src/org/geworkbench/builtin/projects/util/CaARRAYPanel.java: In previous code, layout setting for CaARRAYPanel was omitted, so default FlowLayout was used; then components were added to BorderLayout.CENTER of this FlowLayout, which was a mistake. In jbInit(), set CaARRAYPanel layout to BorderLayout, so that jPanel6 can be added to BorderLayout.CENTER, allowing the CaArrayPanel in JDialog to be resizable.

git-svn-id: https://ncisvn.nci.nih.gov/svn/geworkbench/trunk/geworkbench@5835 8b29bb6e-589a-435a-9a12-3ab907be3462